refactor: splitting index.js
