+++ src/notes/all.js
@@ -1,0 +1,3 @@
+const privnote = require("./websites/privnote");
+
+module.exports = { privnote };
